Move recordEventOnce to platform
