+++ src/applications/disability-benefits/all-claims/content/incidentDate.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AdditionalInfo from '@department-of-veterans-affairs/formation-react/AdditionalInfo';
-import { recordEventOnce } from '../utils';
+import { recordEventOnce } from 'platform/monitoring/record-event';
 import {
   ANALYTICS_EVENTS,
   HELP_TEXT_CLICKED_EVENT,
