+++ src/applications/disability-benefits/all-claims/content/ptsdWalkthroughChoice.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import AdditionalInfo from '@department-of-veterans-affairs/formation-react/AdditionalInfo';
 import { getPtsdClassification } from './ptsdClassification';
-import { recordEventOnce } from '../utils';
+import { recordEventOnce } from 'platform/monitoring/record-event';
 import {
   ANALYTICS_EVENTS,
   HELP_TEXT_CLICKED_EVENT,
