+++ src/applications/disability-benefits/all-claims/content/unemployabilityDisabilities.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import AdditionalInfo from '@department-of-veterans-affairs/formation-react/AdditionalInfo';
 
-import { recordEventOnce } from '../utils';
+import { recordEventOnce } from 'platform/monitoring/record-event';
 
 const helpClicked = () =>
   recordEventOnce({
