+++ src/applications/disability-benefits/all-claims/tests/utils.unit.spec.jsx
@@ -1,7 +1,6 @@
 import sinon from 'sinon';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
-import _ from '../../../../platform/utilities/data';
 
 import {
   makeSchemaForNewDisabilities,
@@ -29,7 +28,6 @@ import {
   ReservesGuardDescription,
   servedAfter911,
   viewifyFields,
-  recordEventOnce,
 } from '../utils.jsx';
 
 describe('526 helpers', () => {
@@ -812,42 +810,6 @@ describe('isAnswering781aQuestions', () => {
         },
       };
       expect(hasHospitalCare(formData)).to.be.false;
-    });
-  });
-
-  describe('recordEventOnce', () => {
-    beforeEach(() => {
-      window.oldDataLayer = _.cloneDeep(window.dataLayer);
-      window.dataLayer = [];
-    });
-
-    afterEach(() => {
-      window.dataLayer = _.cloneDeep(window.oldDataLayer);
-      delete window.oldDataLayer;
-    });
-
-    const testKey = 'help-text-label';
-    const testEvent = {
-      event: 'test-event',
-      [testKey]: 'Test Event',
-    };
-
-    it('should record event if not already in dataLayer', () => {
-      // sanity check to ensure that setup worked
-      expect(window.dataLayer.length).to.equal(0);
-
-      recordEventOnce(testEvent, testKey);
-      expect(window.dataLayer.length).to.equal(1);
-    });
-
-    it('should not record duplicate events', () => {
-      // sanity check to ensure that setup worked
-      expect(window.dataLayer.length).to.equal(0);
-
-      recordEventOnce(testEvent, testKey);
-      recordEventOnce(testEvent, testKey);
-
-      expect(window.dataLayer.length).to.equal(1);
     });
   });
 });
